# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Estoque - Rede Parcerias",
        "description": "Documentação com JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas as categorias",
                "operationId": "9a2f4766d4351f0fa60e4f534a98c8ab",
                "responses": {
                    "200": {
                        "description": "Categorias retornadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Criar nova categoria",
                "operationId": "edf626fe51dae7c25879f1b6b1efbe0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar todos os produtos",
                "operationId": "8c7d00fe13a8e661ad5a967dedf826bd",
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Criar um novo produto",
                "operationId": "15899357c519f76d16796e983cf56735",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Obter detalhes de uma categoria",
                "operationId": "fcd0befe6827f1a7ef7165cf7e9ae32a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualizar uma categoria",
                "operationId": "e06de54d84bb6d3f255067fcb3b13aea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Excluir uma categoria",
                "operationId": "252718d940c9077dd220407da2f5f490",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria excluída com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibir detalhes de um produto",
                "operationId": "f913fa5fc328652556b2ea0a3886441f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar um produto",
                "operationId": "9c295b0312dee0a4bf887833f8420dca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Excluir um produto",
                "operationId": "755d2a11b6f2cf1b06003ae3c76a830c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto excluído com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Categoria",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Papelaria"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Produto",
                "required": [
                    "name",
                    "price",
                    "sku",
                    "quantity",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Caneta azul"
                    },
                    "description": {
                        "type": "string",
                        "example": "Caneta esferográfica azul com ponta fina"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 3.5
                    },
                    "sku": {
                        "type": "string",
                        "example": "CAN-AZUL-001"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-15T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Categorias",
            "description": "Gerenciamento de categorias de produtos"
        },
        {
            "name": "Produtos",
            "description": "Gerenciamento de produtos"
        }
    ]
}